Modified CSS & JS to use 'body' background image
Allowing for graceful improvement, the JS now gets the image URL from
the CSS rule for the '<body>' tag, cleans it up, and uses it.
Also, the screen width threshold for no longer doing the backstretch has
been turned into a variable allow for it to be more easily tweaked.